requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
asttokens==2.0.5
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
bcrypt==3.2.2
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.2
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
entrypoints==0.3
executing==0.8.3
fastapi==0.76.0
geographiclib==1.52
geopy==2.2.0
gnureadline==8.0.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.14.0
httptools==0.4.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.7.1
incremental==16.10.1
instagram-private-api==1.6.0.0
ipykernel==6.9.1
ipython==8.1.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jose==1.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
more-itertools==4.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.6.8
parso==0.8.3
passlib==1.7.4
pexpect==4.6.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.24.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.36
ssh-import-id==5.10
stack-data==0.2.0
starlette==0.18.0
systemd-python==234
tornado==6.1
traitlets==5.1.1
Twisted==18.9.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.7
ufw==0.36
ujson==5.2.0
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.17.6
uvloop==0.16.0
wadllib==1.3.3
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
zipp==3.8.0
zope.interface==4.7.1
